inference-model-cloud-raspberry4/requirements.txt
numpy==1.26.4
pillow==10.3.0
tflite-runtime==2.14.0
fastapi==0.111.0
uvicorn==0.30.1